plugins/claude-design/skills/claude-design/tools/requirements.txt
python-pptx>=0.6.23
playwright>=1.40.0
Pillow>=10.0.0
pymupdf>=1.23.0
lxml>=4.9.0
